// frontend/src/lib/types.ts
export interface ApiResponse<T = any> {
  status: "success" | "error";
  message?: string;
  data?: T;
  group_id?: string;
  token?: string; // Optional token for login/signup responses
  // error?: string // Optional error message
}
// ...
export interface StartCreateGroupPayload {
  livestock_id: string;
  groupName: string;
  description: string;
  totalSlot: number;
  slotPrice: number;
  // creatorInitialSlots: number;
  slotTaken: number;
}

export interface CompleteCreateGroupPayload {
  groupId: string;
  paymentMethod: "wallet" | "others";
  paymentReference?: string;
  amount: string;
}

export interface CreateGroupResponse {
  group_id: string;
}
// ...
export interface Livestock {
  id?: number;
  livestock_id: string;
  name: string;
  breed?: string;
  weight?: number;
  price: number;
  imageUrl?: string;
  description?: string;
  available: boolean;
  minimum_amount: number; // Updated to match model
  createdAt: string;
  updatedAt: string;
}